Add News command
A basic command that works like /roomintro, but doesn't show up when a user joins a room; "/news" to view and "/news (text)" to edit. Since I haven't fully gotten the hang of writing commands please be careful—"/news     " will cause the news to be changed to a bunch of spaces, and it does not return the source html. I will work on this more but it's something for now